Replace moment with luxon
For some reason GHA returns by default a UTC time object but when we get the timestamp it return a localized timestamp. I have no idea how to fix this without hardcoding the timezone difference so instead we'll just give up and pretend everything is in CST. Moving to luxon since its newer and cheaper to fetch.